* 'master' of github.com:guardian/workflow: (53 commits)
  put config infront of import to remove race condition
  Nicer icon, support for high res
  Adds favicon
  Rename contributing.md to CONTRIBUTING.md
  contributing file as per Robert Rees's reccomendation
  remove unneeded import
  Changed bg colour of buttons 'New' "Import from composer'
  Changed deadline, created and section font-size
  Changed the side bar link text size
  Fix hover state of filter links
  Add text -uppercase to content-list group heading
  Remove unnecessary sidebar widths
  Refactor click event handling on sidebar__filter
  Remove unused comment
  Sass refactor
  Makes production office time picker fit coloumn
  Removes width attribute on sidebar__heading
  Change delete icon colour on dark background
  Update _content-list.scss
  Update _content-list.scss
  ...

Conflicts:
	prototype/public/components/sidebar-filter/_sidebar-filter.scss
	prototype/public/layouts/dashboard/_sidebar.scss

+++ prototype/public/components/content-list-drawer/content-list-drawer.js
@@ -34,7 +34,6 @@ var wfContentListDrawer = function ($rootScope, config, $timeout, $log, $window,
              * Accessible on $scope for use on 'hide' button in drawer
              */
             this.hide = function () {
-
                 $element.one(transitionEndEvents,() => {
                     $scope.contentList.selectedItem = null;
                     $scope.$apply();
@@ -71,6 +70,8 @@ var wfContentListDrawer = function ($rootScope, config, $timeout, $log, $window,
              * Listen for event triggered by click in external contentItemRow directive to show or hide drawer
              */
             $rootScope.$on('contentItem.select', ($event, contentItem, contentListItemElement) => {
+                $scope.awaitingDeleteConfirmation = false;
+
                 if (contentItem.status === 'Stub') {
                     $scope.$emit('stub:edit', contentItem.item);
                     return;
@@ -172,7 +173,9 @@ var wfContentListDrawer = function ($rootScope, config, $timeout, $log, $window,
             /**
              * Delete manually as no event or tracking yet
              */
+            $scope.awaitingDeleteConfirmation = false;
             $scope.deleteContentItem = function () {
+                if(!$scope.awaitingDeleteConfirmation) { $scope.awaitingDeleteConfirmation = true; return; } 
 
                 contentService.remove($scope.contentItem.id)
                     .then(() => {
